Synthetic time series for event detection

A list of univariate time series for event detection

example1: a single (uncommon value) anomaly example in a stationary time series

example2: a single (common value) anomaly example in a stationary time series

example3: a single anomaly example in a trend time series

example4: a change point in a time series

example5: a change point with anomaly (uncommon value) in a time series

example6: a change point with anomaly (common value) in a time series

example7: a change point with anomaly (uncommon value) in a seasonal time series

example8: a change point with anomaly (common value) in a seasonal time series

example9: multi behavior time series (1-200: stationary, 201-400: trend, 401-600: structural break; 601-800: heteroscedasticity, 801:1000 random walk)

example10: multiple anomalies

example11: stationary time series

example12: trend time series

example13: structural break

example14: heterocedasticity

example15: motif

example16: discord

example17: repetitive anomaly

example18: repetitive anomaly

#'

Usage

data(har_examples)

Format

A list of time series.

Source

Harbinger package

References

Harbinger package

Examples

data(har_examples)
serie <- har_examples$example1